rename notificationsApiTypes file to notificationsApi.types
