+++ src/features/notifications/api/notificationsApi.types.ts
